# docs/swagger.yaml
basePath: /
definitions:
...
  auth.RequestBulkMenuRole:
    properties:
      data:
        items:
          $ref: '#/definitions/auth.RequestMenuRoleItem'
        type: array
      roleId:
        type: string
    required:
    - roleId
    type: object
  auth.RequestMenuFormat:
    properties:
      action:
        type: string
      description:
        type: string
      icon:
        type: string
      level:
        type: integer
      link:
        type: string
      name:
        type: string
      parentId:
        type: string
      permissionLabel:
        type: string
      seq:
        type: integer
    required:
    - link
    - name
    type: object
  auth.RequestMenuRoleItem:
    properties:
      menuId:
        type: string
      permission:
        type: string
    type: object
...
  response.Base:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  transaction.RequestBulkPenilaian:
    properties:
      data:
        items:
          $ref: '#/definitions/transaction.RequestDetailPenilaian'
        type: array
      usulanId:
        example: uuid-usulan
        type: string
    required:
    - data
    - usulanId
    type: object
  transaction.RequestDetailPenilaian:
    properties:
      kriteriaId:
        example: uuid-kriteria
        type: string
      nilaiInput:
        example: 85.5
        type: number
    required:
    - kriteriaId
    - nilaiInput
    type: object
  transaction.RequestHitungTopsis:
    properties:
      tahapVersi:
        description: '''RKP'' atau ''RAPBDes'''
        example: RKP
        type: string
      tahunAnggaran:
        example: 2025
        type: integer
    required:
    - tahapVersi
    - tahunAnggaran
    type: object
  transaction.RequestPaguAnggaran:
    properties:
      jumlahPagu:
        example: 500000000
        type: number
      sumberDanaId:
        example: masukkan-uuid-sumber-dana-disini
        type: string
      tahun:
        example: 2025
        type: integer
    required:
    - jumlahPagu
    - sumberDanaId
    - tahun
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Backend untuk Sistem Pendukung Keputusan (SIPROPEDA) - Metode TOPSIS.
  title: SIPROPEDA API
  version: "1.0"
paths:
...
  /v1/menu:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua Master Menu
      tags:
      - Menus
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Menu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestMenuFormat'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah Master Menu
      tags:
      - Menus
  /v1/menu-role:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID
        in: query
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil Hierarki Menu by Role ID
      tags:
      - Menus Role
  /v1/menu-role/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Mapping
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestBulkMenuRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Pasang Akses Menu ke Role
      tags:
      - Menus Role
  /v1/menu/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Menu
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus Master Menu
      tags:
      - Menus
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Menu
        in: path
        name: id
        required: true
        type: string
      - description: Data Menu
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestMenuFormat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Update Master Menu
      tags:
      - Menus
  /v1/pagu-anggaran:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua data Pagu Anggaran
      tags:
      - Pagu Anggaran
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Pagu Anggaran
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestPaguAnggaran'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah data Pagu Anggaran
      tags:
      - Pagu Anggaran
  /v1/pagu-anggaran/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Pagu Anggaran
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus data Pagu Anggaran (Soft Delete)
      tags:
      - Pagu Anggaran
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Pagu Anggaran
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail Pagu Anggaran by ID
      tags:
      - Pagu Anggaran
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Pagu Anggaran
        in: path
        name: id
        required: true
        type: string
      - description: Data yang akan diedit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestPaguAnggaran'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Update data Pagu Anggaran
      tags:
      - Pagu Anggaran
  /v1/penilaian-usulan:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Penilaian
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestBulkPenilaian'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Simpan Penilaian Usulan (Bulk)
      tags:
      - Penilaian Usulan
  /v1/penilaian-usulan/{usulanId}:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Usulan Proyek
        in: path
        name: usulanId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil Nilai berdasarkan Usulan ID
      tags:
      - Penilaian Usulan
  /v1/perankingan/arsip:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tahun Anggaran
        in: query
        name: tahun
        required: true
        type: integer
      - description: 'Tahapan (misal: RKP)'
        in: query
        name: tahap
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil Arsip Perankingan
      tags:
      - Perankingan
  /v1/perankingan/hitung:
    post:
      consumes:
      - application/json
      description: Endpoint untuk menghitung perangkingan menggunakan metode TOPSIS
        berdasarkan tahun dan tahap
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Parameter Perhitungan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestHitungTopsis'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Kalkulasi Ranking TOPSIS
      tags:
      - Perankingan
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"